A Tournament can now generate its own matches. This was again extracted as a module.  Modules get more powerful as they get more generic.  This is far from generic.
